Rename concrete repo to eloquent habit repo
