+++ api/src/HabitTracking/Infrastructure/Registration.php
@@ -3,7 +3,6 @@
 namespace HabitTracking\Infrastructure;
 
 use HabitTracking\Domain\Contracts\HabitRepository;
-use HabitTracking\Infrastructure\Eloquent\HabitRepository as EloquentHabitRepository;
 use Illuminate\Support\ServiceProvider;
 
 class Registration extends ServiceProvider
