+++ api/tests/Integration/EloquentHabitRepositoryTest.php
@@ -4,7 +4,7 @@ use App\Models\User;
 use HabitTracking\Domain\Exceptions\HabitNotFound;
 use HabitTracking\Domain\Habit;
 use HabitTracking\Domain\HabitId;
-use HabitTracking\Infrastructure\Eloquent\HabitRepository as EloquentHabitRepository;
+use HabitTracking\Infrastructure\EloquentHabitRepository;
 use HabitTracking\Presentation\Habit as EloquentHabit;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\Support\HabitModelFactory;
